Implement GPMO_backtracking for arbitrary vectors
A new function GPMO_ArbVec_backtracking has been added to
permanent_magnet_optimization.cpp. It is an implementation of the
GPMO backtracking (GPMOb) algorithm that allows the user to specify
lists of arbitray allowable unit-length polarization vectors for each
dipole in the grid.

The user interface for GPMO_ArbVec_backtracking differs from that of
GPMO_backtracking in the following ways:

  - The user must provide the list of allowable polarization vectors
    via the argument `pol_vectors` (as for GPMO_Arbvec)

  - The argument `K` effectively stands for the maximum number of
    iterations to run for the optimization (this will be greater than
    or equal to the number of magnets ultimately placed)

  - The argument `thresh_angle` specifies the threshold value of the
    angle in radians between the polarization vectors of two magnets
    being compared during the backtracking procedure, above which the
    pair would be removed. For the original GPMO_backtracking in the
    Cartesian basis, this value is effectively pi.

  - The argument `max_nMagnets` specifies the maximum number of magnets
    to place in the solution. (The optimization will terminate if the
    number of magnets reaches this value.)

In this commit, GPMO_ArbVec_backtracking has been shown to produce
identical iterations to GPMO_backtracking in the Cartesian basis if
the allowable vectors supplied to GPMO_ArbVec_backtracking are the
Cartesian unit vectors.