Fix memory overwrite in GPMOb procedures
Edits are made to GPMO_backtracking and GPMO_ArbVec_backtracking
(permanent_magnet_optimization.cpp) that write data to the
num_nonzeros array. The index is changed from print_iter to
print_iter-1 to account for the fact that print_iter is incremented
in the call to print_GPMO which occurs just before the write
to num_nonzeros.

Separately, a ValueError condition is added to the function GPMO
(permanent_magnet_optimization.py) to prevent the usage of the
GPMO_ArbVec_backtracking algorithm with PermanentMagnetGrid class
instances that do not use the Cartesian basis. Other bases are
currently not supported by GPMO_ArbVec_backtracking. This capability
could be added in the future if it is considered worthwhile. However,
such a change would not expand the optimization capabilities of
GPMO_ArbVec_backtracking, as the current version already allows
polarization vectors in any direction to be specified for each magnet.