Change GPMO backtracking order for arb. vectors
The GPMOb algorithm for arbitrary polarization vectors is edited such
that the backtracking procedure iterates through magnets in the order
in which they appear in the solution vector, rather than the order
in which they were switched on for the solution. This change was made
to facilitate the option (to be added later) of having the user provide
an initial guess for the solution, in which case the switch-on order
would be undefined.