Enable initialization of GPMOb (arb. vectors)
The function GPMO_ArbVec_backtracking now takes an additional argument,
`x_init`, containing initial (normalized) values of the dipole moments
at each grid point. A new function, initialize_GPMO_ArbVec, is called
to populate the solution array and various helper data structures used
during the optimization according to `x_init`. If a normalized dipole
moment supplied in `x_init` does not match one of the allowable
normalized moments for the corresponding magnet in `pol_vectors`,
the dipole moment will be initialized to the closest allowable
vector or zero according to the Euclidean distance to the vector in
`x_init`.

When using the Python wrapper function `GPMO` with the algorithm
"ArbVec_backtracking", the initialization is performed by supplying a
2D array with the dipole moments (nominal values; *not* normalized) as
an entry in the `algorithm_kwargs` dictionary with the key "m_init".
If this entry is not supplied, the wrapper function will initialize
all magnets to zero by default, matching the previous behavior.

Two further modifications are made to GPMO_ArbVec_backtracking to
accommodate the initialization. First, the backtracking loop is
activated after the first iteration (when k=0). Previously,
backtracking was not invoked until k=nBacktracking. This change should
not alter the outcome of the optimization in the case of default
initialization. In addition, the print_iter function is invoked
before any GPMO iteration is taken in order to have a record of
the initialization in the m_history file.