Fix bug in CI test and adjust GPMO magnet counting
This commit fixes a bug in one of the tests in the CI workflow and
makes a small change to the point in the command sequence of
GPMO_backtracking where magnets are counted.

The bug in the CI workflow arose from comparisons of iteration history
data obtained from different GPMO algorithms in the test script
tests/solve/test_pm_optimization.py. Specifically, the function
GPMO_ArbVec_backtracking (src/simsoptpp/
permanent_magnet_optimization.cpp) outputs an extra history entry for
the user-input initialized solution. Therefore, when raw output
history arrays from GPMO_ArbVec_backtracking were compared with
corresponding history arrays from other GPMO procedures, an array
dimension mismatch error would occur. To address this issue, the test
procedure now ignores the first entries of the history arrays
from GPMO_ArbVec_backtracking, which correspond to the data for the
initialization.

In addition, a small change is made to the command sequence in the
outer loop of GPMO_backtracking
(src/simsoptpp/permanent_magnet_optimization.cpp). Previously,
the incremented quantity num_nonzero was incremented right after
data for the corresponding iteration was saved and printed to screen.
In the present version, num_nonzero is updated first, so that the
printed data will correctly reflect the number of magnets for its
iteration.

In addition, some edits have been made to address issues that arose
in the linting workflow.