Implement matrix of B-normal contributions
The WireframeField class now contains methods for generating a matrix
relating each wireframe segment to its contribution to the normal
component of the magnetic field on the plasma boundary. For each
segment, the matrix includes contributions from the equivalent
segments in each half period, each of which are constrained to
carry an identical current when assuming stellarator symmetry.

The matrix will be useful for certain optimizations meant to limit the
normal field on the plasma boundary, including the linear least-squares
approach as well as the greedy approach.

This is implented through two methods in the WireframeField class:

  - dB_by_dsegmentcurrents: calculates the contribution of each segment
    to the magnetic field vector, per unit current, at the reference
    points

  - dBnormal_by_dsegmentcurrents_matrix: calculates the contribution of
    each segment to the normal component of the magnetic field on a
    user input toroidal surface. The method above is called internally.
    Note that the reference points in the instance of the
    WireframeField class are assumed to be the same as the quadrature
    points of the input toroidal surface.