Add PortSet and CircularPort classes
Two classes are implemented in ports.py to aid in the design of
magnet solutions that avoid collisions with ports. PortSet represents
a generic set of ports to avoid. Its key method is collides(), which
determines whether a set of user-provided input points collide with
any of the member ports. Each of the member ports are assumed to be
class instances with their own collides() function implemented.

Currently the only type of port for which a model is available is a
simple cylindrical port with a circular cross-section. This is
implemented in the CircularPort class.