Enable generic spatial constraints for wireframes
The ToroidalWireframe class now has a method
constrain_colliding_segments(), which constrains segments on which
test points are shown to collide with external objects or otherwise
violate spatial constraints. The method takes as an argument a user-
supplied function that determines whether input test points are
colliding; for example, the method collides() of the PortSet class
(ports.py).