Generalize wrapper for RCLS and GSCO
A new wrapper function, optimize_wireframe, serves as a high-level
wrapper for wireframe optimization. It can be used to perform both
a GSCO and an RCLS solve. Parameters specific to the optimizer are
supplied in a dictionary. In addition to handling field calculations
if necessary, optimize_wireframe calls lower-level wrappers specific
to each optimizer: rcls_wireframe and gsco_wireframe.

In addition, the calculation of the inductance matrix is factored out
as the function bnorm_obj_matrices, which can be called separately
by the user.