Add unit test for port classes
A new file, test_ports.py, is added with unit tests for the port
classes in ports.py. Most of the tests entail verifying that sets of
test points lie inside and outside of sets of circular and rectangular
ports where expected.

In the process of implementing these tests, some changes were made to
the CircularPort and RectangularPort classes to fix a bug in the
implementation of stellarator symmetry:

  - colliding() methods are generalized to correctly assess collisions
    irrespective of whether l0 < l1 or l0 > l1

  - repeat_via_symmetries() methods now negate the parameters l0 and l1
    for transformations to the alternate half-period

  - Related to the above, l0 and l1 switch places in addition to being
    negated when transforming to the alternate half-period, although
    this is not strictly necessary in light of the change to the
    colliding() methods listed above