Change sign convention for target normal field
In the functions optimize_wireframe and bnorm_obj_matrices, the sign
convention for specifying the target normal component of the magnetic
field on the plasma boundary has changed. Previously, the quantity
to be supplied, which had the parameter name `bn_plas_curr`, was the
negative of the target normal field, i.e. the normal field produced by
internal plasma currents. In the present version, the quantity to be
supplied has the name `bnorm_target` and directly corresponds to the
normal field that the wireframe (and external magnetic fields if
applicable) must produce.

This updated sign convention makes the wireframe optimization functions
consistent with other Simsopt features such as the SquaredFlux
objective.

NOTE: this change is not backward compatible. Scripts that relied on
the previous convention will fail because the input parameter
`bn_plas_curr` is no longer accepted.