Fix wireframe optimization test with target bnormal
The unit tests that supply a target normal field to
`optimize_wireframe` via the `bnorm_target` keyword argument
have been modified to use a different target bnormal distribution.
In the present version, the target bnormal is designed to arise
from an external uniform vertical magnetic field. With this
distribution, the tests are behaving as expected.

The previous version supplied a target bnormal that was uniform across
the plasma boundary, which would have implied a non-divergence-free
magnetic field; hence, it is not clear how the wireframe solutions
would have differed in response to changes in the magnitude of this
target distribution.

Also, as a workaround to avoid introducing an `area_weight` attribute
to the wireframe class instance during the optimization procedure,
the tests set the `area_weighted` keyword argument to False; this way,
it is not necessary to use the area weights for the quantitative
comparisons in the unit tests.