Update instructions for recompiling simsoptpp
The previous version of the instructions for recompiling simsoptpp
presumed that the shared object file for simsoptpp used by Python
would be stored in the ``src`` directory; however, this is no longer
the case in the latest versions.

This commit updates the instructions to be agnostic to the location
of the shared object file referenced by Python, using the Python
command "print(simsoptpp.__file__)" to directly look up the correct
location.