Add cmake flag to compilation instructions
It may be necessary to add a flag to the ``cmake`` command to inform
cmake of the location of the pybind11 installation. The present
version of the instructions insures that pybind11 is installed in
the virtual environment and then constructs the relevant path
from info gained from the __path__ property of pybind11 when imported
in python.