Put wireframe half-periods in toroidal spatial order
The coordinates of the nodes of a wireframe are stored in a list of
arrays, where each array corresponds to one half-period. This commit
adjusts the order in which the half-period arrays appear in the list.
In the current version, successive arrays correspond to successive
half-periods in toroidal order. Previously, these arrays were not in
toroidal order. This changes is meant to streamline algorithms for
coil finding.