Add method to find coils in a wireframe
The method `find_coils` is added to the ToroidalWireframe class. This
method performs a graph search to identify distinct, closed paths
of current within the wireframe and categorize them into groups
with equivalent geometry and currents.

An additional method `has_current_crossings` to check for crossed
current paths in the wireframe. This is helpful as a checker for
`find_coils`, which doesn't work in distributions with crossed paths;
it can also be useful in other contexts.

A set of unit tests are added to test_wireframe_toroidal.py for the
new features.