Add function to extract Simsopt coils from a wireframe
The function `coils_from_wireframe` is added to wireframefield.py that
returns simsopt `Coil` class instances for every coil in a wireframe
whose current distribution has no crossings. The function first calls
the `find_coils` method of a ToroidalWireframe class instance to
identify and group the coils. Then, for each coil, the function
instantiates a `Coil` or `RotatedCoil` class instance in which the
curve is a `CurveXYZFourier` fitted to the node points of each coil.

A set of corresponding unit tests for this function is added to
test_wireframefield.py.