Increase flexibility of helical coil interface
The `add_helical_coil_currents` method of the ToroidalWireframe class
is updated to allow for greater flexibility in constructing helical
coils. Now, the user can specify arbitrary crossing points for helical
current paths at each symmetry plane, although the method will raise
an error if the configuration violates current continuity (and,
implicitly, stellarator symmetry).

The interface is therfore incompatible with the previous commit.

In addition, some unit tests are added for wireframe field
calculations, and the tests for the `add_helical_coil_currents` method
in test_wireframe_toroidal.py are updated for the new interface and
made more comprehensive.