Initialize index array as integer type
The array meant to store indices in the `make_plot_2d` method of
the ToroidalWireframe class is now initialized explicitly as an
integer array. The aim is to prevent a bug from occurring in
some tests with one of these arrays that appears to be due to a
floating point error.